requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
asttokens==3.0.0
attrs==25.4.0
blake3==1.0.8
cachetools==6.2.1
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
compressed-tensors==0.11.0
debugpy==1.8.17
decorator==5.2.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
easydict==1.13
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.120.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.20.0
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
interegular==0.3.3
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
lark==1.2.2
lm-format-enforcer==0.11.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral_common==1.8.5
mpmath==1.3.0
msgspec==0.19.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
ollama==0.6.0
openai==2.6.1
openai-harmony==0.0.4
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
outlines_core==0.2.11
packaging==25.0
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pdf2image==1.17.0
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.15.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rpds-py==0.28.0
safetensors==0.6.2
scipy==1.16.2
sentencepiece==0.2.1
sentry-sdk==3.0.0a7
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.20.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
vllm==0.11.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
zipp==3.23.0
